walletfreak/requirements.txt
django>=5.0
firebase-admin>=6.2.0
gunicorn>=21.2.0
requests
pillow
python-dateutil
markdown>=3.9
whitenoise
python-dotenv
django-ninja>=1.0
stripe
certifi
psycopg2-binary
dj-database-url

selenium
webdriver-manager
xai-sdk
beautifulsoup4
playwright
playwright-stealth
